requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes
annotated-doc==0.0.4
    # via typer
anyio==4.12.1
    # via httpx
blob==0.16
    # via quyet
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
click==8.3.1
    # via typer
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   tqdm
contourpy==1.3.3
    # via matplotlib
cuda-bindings==12.9.4 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
cuda-pathfinder==1.3.4 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via cuda-bindings
cycler==0.12.1
    # via matplotlib
dataclasses==0.8
    # via quyet
filelock==3.21.2
    # via
    #   huggingface-hub
    #   torch
fonttools==4.61.1
    # via matplotlib
fsspec==2026.2.0
    # via
    #   huggingface-hub
    #   torch
h11==0.16.0
    # via httpcore
hf-xet==1.2.0 ; platform_machine == 'AMD64' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via huggingface-hub
huggingface-hub==1.4.1
    # via timm
idna==3.11
    # via
    #   anyio
    #   httpx
jinja2==3.1.6
    # via torch
kiwisolver==1.4.9
    # via matplotlib
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
matplotlib==3.10.8
    # via quyet
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.6.1
    # via torch
numpy==2.4.2
    # via
    #   contourpy
    #   matplotlib
    #   opencv-python
    #   pycocotools
    #   quyet
    #   torchvision
nvidia-cublas-cu12==12.8.4.1 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cuda-nvrtc-cu12==12.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cuda-runtime-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cudnn-cu12==9.10.2.21 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cufft-cu12==11.3.3.83 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cufile-cu12==1.13.1.3 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-curand-cu12==10.3.9.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cusolver-cu12==11.7.3.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cusparse-cu12==12.5.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.7.1 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-nccl-cu12==2.27.5 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-nvjitlink-cu12==12.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvshmem-cu12==3.4.5 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-nvtx-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
opencv-python==4.13.0.92
    # via quyet
packaging==26.0
    # via
    #   huggingface-hub
    #   matplotlib
pillow==12.1.1
    # via
    #   matplotlib
    #   quyet
    #   torchvision
pycocotools==2.0.11
    # via quyet
pygments==2.19.2
    # via rich
pyparsing==3.3.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   timm
rich==14.3.2
    # via typer
safetensors==0.7.0
    # via timm
setuptools==82.0.0
    # via torch
shellingham==1.5.4
    # via
    #   huggingface-hub
    #   typer
six==1.17.0
    # via python-dateutil
sympy==1.14.0
    # via torch
timm==1.0.24
    # via quyet
torch==2.10.0
    # via
    #   quyet
    #   timm
    #   torchvision
torchvision==0.25.0
    # via
    #   quyet
    #   timm
tqdm==4.67.3
    # via
    #   huggingface-hub
    #   quyet
triton==3.6.0 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
typer==0.23.1
    # via typer-slim
typer-slim==0.23.1
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   anyio
    #   huggingface-hub
    #   torch
